docs/getting-started: document absolute path for config bind mount
Document using an absolute path to `your_config.fcc`. If an
absolute path isn't supplied, `podman` interprets this as the
name for a volume, and mounts the volume at `/config.fcc` in the
container. With `/path/to`, it is clearer from the error message
what the problem in the invoked command is if left unchanged.

```
$ podman run --rm -v your_config.fcc:/config.fcc:z quay.io/coreos/fcct:v0.2.0 -pretty -strict -input /config.fcc > transpiled_config.ign
failed to read /config.fcc: read /config.fcc: is a directory

$ podman run --rm -v /path/to/your_config.fcc:/config.fcc:z quay.io/coreos/fcct:v0.2.0 -pretty -strict -input /config.fcc > transpiled_config.ign
Error: error checking path "/path/to/your_config.fcc": stat /path/to/your_config.fcc: no such file or directory
```