base: warn if file/dir modes appear to have been specified in decimal
Users sometimes forget to write file/directory modes with a leading
zero and end up with a nonsensical configuration.  If a specified mode
is unreasonable but would have been reasonable if it had been written
in octal, report a warning.  Apply to all spec versions because it's
just a warning.