config/fcos/v1_[34]: un-RAID ESP
It's not safe to RAID the ESP because the firmware might decide to write
to an individual replica, causing the RAID to desynchronize.  Instead,
create a separate filesystem for each replica.

Don't address backward compatibility because we're making this change
soon after the functionality landed, and before it was documented in the
OS.