config/fcos: Avoid merging invalid configs
When the boot_device mirror functionality was added we added
`ToIgn3_XUnvalidated` functions to the `fcos` `v1_3` & `v1_4_exp`
configs which were not checking that the base configs were valid before
attempting to merge in child configs. Add a quick check that the report
for the base config is not fatal before continuing.