base/*: add translations for list bases added by trees
storage.trees currently can't be ported to use config merging without
introducing regressions.  Therefore, it still works with the live
TranslationSet, and must be careful to add translations only for lists
not already included in the set.

Existing tests didn't catch this because the successful ones all included
overrides, and the overrides were generating the missing translations.