config/openshift/*: error on btrfs filesystem format
RHCOS doesn't ship mkfs.btrfs, so any attempt to format a btrfs filesystem
will fail.  Detect this before emitting the config.  Since this validation
doesn't impose a new restriction on the config, add it to the 4.8.0 stable
spec as well.

I opted to put the new validator in a separate function (since a different
team and codebase is responsible for the limitation) but to test it in
the same test function (since there's no user-visible or technical
distinction between the validations within the Butane codebase).

There's no current requirement to do this validation in JSON space, but
it's consistent with existing code, and we might eventually have sugar
that allows the user to specify the filesystem type.