openshift/v4_9_exp: drop Ignition KernelArguments field
We don't want to expose KernelArguments in two different places that
have different semantics. Drop the Ignition KernelArguments fields to
encourage the use of the MCO's related field as it will also affect
subsequent boots.

In addition, the MCC in OCP 4.9 is unable to parse Ignition 3.3.0 configs
that are not also valid 3.2.0 configs.  Document that the Ignition
KernelArguments fields are not supported by the MCC.

Co-authored-by: Stephen Lowrie <slowrie@redhat.com>