config: drop TranslateBytesOptions.Strict
The only thing it did was cause util.TranslateBytes() to instruct the YAML
decoder to fail on keys that aren't in the schema.  This isn't necessary
because util.TranslateBytes() also calls Ignition's ValidateUnusedKeys(),
which detects the same condition and adds warnings to the report.  The
caller can then fail on warnings if it wishes.

The flag was originally added in 6d57f1f3f155.  The ValidateUnusedKeys
call was added shortly thereafter, in b4d04b2ef72c.

This is an API change.