docs/config-*: document result of setting `systemd.units.mask` false
This was documented upstream in coreos/ignition@18ac09ab0f but never
made it downstream.  The behavior was added in Ignition 2.8.0 and is
always available in Ignition specs 3.3.0+.  Don't bother adding a
versioning note to the docs: flatcar and r4e aren't affected at all,
and no version of FCOS or RHCOS supported by Butane has an old enough
Ignition to not have the behavior.