config/openshift: attach missing label error to labels map
The missing label may or may not exist, depending whether the value is
"" or actually absent.  If it doesn't exist, we shouldn't include a path
to it in the report.

This shouldn't actually affect the output, since the line/column
correlator would have fallen back to the parent field anyway.