tests: verify that report paths are valid
A common mistake is to create ContextPaths that aren't valid references to
an item in the config.  DebugVerifyCoverage() should catch such paths in
a TranslationSet, but validation errors (during either the validation or
translation passes) may add paths to a report without a corresponding
TranslationSet entry.  Add helper functions to validate paths in a report,
and use them in every test case.

In particular, call the helpers even in tests that expect empty reports,
to reduce the risk that a test created in the future will omit the check.