fcos/v1_6_exp: warn if config attempts to reuse partition by label
Ignition matches existing partitions by number, not by label.  When
reusing an existing partition table, if the config specifies a label
and leaves the number at the default of 0, Ignition will try to add a
new partition rather than reusing any existing one.  That was probably
not the user's intention, so warn in this case.

However, the boot disk is a common exception, since it has a fresh
partition table on every run.  On FCOS we can detect the alias
/dev/disk/by-id/coreos-boot-disk and not warn, but that alias was added
recently and older configs won't use it.  On other variants we don't
have any way to detect the boot disk.  Therefore, don't implement the
warning in base and don't make it retroactive to existing stable specs;
only add it to the FCOS experimental spec.