Stop requiring translation functions to return yaml paths
We were expecting this but weren't ensuring it.  The r4e
checkForForbiddenFields() functions weren't, causing their errors not to
identify the correct line and column.

Since 94829f00fd, TranslateReportPaths() automatically ignores paths that
are already in yaml syntax, so we can just call it once from
config/util.Translate() and stop requiring each individual translator to
call it.  The exception is the CheckForElidedFields() logic in the
openshift spec ToIgn*Unvalidated() functions, since the caller can't do
it for us: the returned report needs to be translated using entries that
are elided from the returned TranslationSet.

This is a breaking change for any external caller using the *Unvalidated
functions, since those may now (legally) return a mix of yaml and json
paths.